Se cambia tipo de variable a Long

+++ src/main/java/com/home_rental_solution/ms_propiedades/client/AnfitrionClient.java
@@ -10,5 +10,5 @@ public interface AnfitrionClient {
     //GET /anfitriones/id/validar
     //devuelve true si el anfitrion esta verificado
     @GetMapping("/api/v1/anfitriones/{id}/validar")
-    boolean validarAnfitrion(@PathVariable int id);
+    boolean validarAnfitrion(@PathVariable Long id);
 }
